+++ Block/Adminhtml/Button/Reset.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Snowdog\Menu\Block\Adminhtml\Button;
+
+use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
+
+class Reset extends AbstractButton implements ButtonProviderInterface
+{
+    public function getButtonData(): array
+    {
+        return [
+            'label' => __('Reset'),
+            'on_click' => 'window.location.reload();',
+            'class' => 'reset',
+            'sort_order' => 30
+        ];
+    }
+}