+++ Block/Adminhtml/Button/Save.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Snowdog\Menu\Block\Adminhtml\Button;
+
+use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
+
+class Save extends AbstractButton implements ButtonProviderInterface
+{
+    public function getButtonData(): array
+    {
+        return [
+            'label' => __('Save'),
+            'class' => 'save primary',
+            'data+attribute' => [
+                'mage-init' => ['button' => ['event' => 'save']],
+                'form-role' => 'save'
+            ],
+            'sort_order' => 50
+        ];
+    }
+}